[ALPS04346378] Ion: sync structure update
Ion cache sync structure update

MTK-Commit-Id: 6ff91a2805ec05d995f4c83ab5c556cdf10eb0ee

Feature: [Module]ION/M4U
CR-Id: ALPS04346378

+++ drivers/staging/android/ion/mtk/ion_drv.h
@@ -103,7 +103,7 @@ struct ion_sys_cache_sync_param {
 	void *va;
 	unsigned int size;
 	enum ION_CACHE_SYNC_TYPE sync_type;
-	unsigned long iova;
+	unsigned long long iova;
 };
 
 enum ION_DMA_TYPE {
